Refactors views/fusionaccess
